refactor: Remove adapter template directive
