+++ ui/src/app/connect/components/adapter-configuration/adapter-configuration.component.ts
@@ -19,10 +19,7 @@
 import { Component, Input, OnInit, ViewChild } from '@angular/core';
 import { UntypedFormBuilder, UntypedFormGroup } from '@angular/forms';
 import { MatStepper } from '@angular/material/stepper';
-import {
-    AdapterDescription,
-    LinkageData,
-} from '@streampipes/platform-services';
+import { AdapterDescription } from '@streampipes/platform-services';
 import { ShepherdService } from '../../../services/tour/shepherd.service';
 import { EventSchemaComponent } from './schema-editor/event-schema/event-schema.component';
 import { TransformationRuleService } from '../../services/transformation-rule.service';
@@ -45,7 +42,6 @@ export class AdapterConfigurationComponent implements OnInit {
     @Input() adapter: AdapterDescription;
     @Input() isEditMode;
 
-    linkageData: LinkageData[];
     myStepper: MatStepper;
     parentForm: UntypedFormGroup;
     pageTitle = '';
