+++ ui/src/app/connect/components/adapter-configuration/adapter-settings/configuration-group/configuration-group.component.scss
